🧹 Clean up helpers part 0

+++ app/lib/cms/semantic_form_builder.rb
@@ -15,17 +15,13 @@ module CMS
     end
 
     def delete_button
+      return disabled_delete_button unless object_can_be_destroyed?
+
       template.link_to('Delete', template.url_for(action: :destroy),
                        'data-method': :delete,
                        class: 'delete pf-c-button pf-m-danger',
                        title: "Delete this #{@object.class.model_name.human.downcase}",
                        'data-confirm': "Do you really want to delete this #{@object.class.model_name.human.downcase}?")
-    end
-
-    def disabled_delete_button
-      template.link_to('Delete', '',
-                       class: 'delete pf-c-button pf-m-danger pf-m-disabled',
-                       title: "You cannot delete this #{@object.class.model_name.human.downcase} because it's being used")
     end
 
     def hide_button(*args)
@@ -44,6 +40,16 @@ module CMS
     def actions(*args, &block)
       # HACK: CMS design is too far away from the rest of the app so skip SemanticFormBuilder implementation
       ::Formtastic::FormBuilder.instance_method(:actions).bind(self).call(*args, &block)
+    end
+
+    private
+
+    def object_can_be_destroyed?
+      object.try(:can_be_destroyed?) != false
+    end
+
+    def disabled_delete_button
+      template.link_to('Delete', '', class: 'delete pf-c-button pf-m-danger pf-m-disabled')
     end
   end
 end
